Add Github workflow
The beauty of `nob` is how we can build Musialiser with a simple
interface:

```
cc -o nob nob.c
./nob
```

(or equivalent in Windows)

We can leverage this simplicity and slap a simple Github workflow to
test the build in all platforms. (Feel free to insert the flex slap meme
here).

I took the liberty of setting this to be triggered by `push` and
`pull_request`, so it can do a basic check whenever someone wants to
create a PR in this repo.

I _think_ this can be extended to test build for multiple configurations
(hotreload, etc) but I decided to start with something simpler.